requirements.txt
asgiref==3.7.2
Django==4.2.4
django-sortedm2m==3.1.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
Pillow==10.0.0
requests==2.31.0
psycopg2-binary==2.9.3 